pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "wheeler"
version = "0.1.0"
description = "A CLI research assistant for scientists, powered by Claude"
requires-python = ">=3.10"
dependencies = [
    "claude-agent-sdk",
    "typer>=0.9",
    "rich>=13.0",
    "neo4j>=5.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "prompt-toolkit>=3.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
wheeler = "wheeler.cli:app"
wheeler-tools = "wheeler.tools.cli:app"

[tool.setuptools.packages.find]
include = ["wheeler*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
